requirements.txt
fastapi
uvicorn
pymongo
python-multipart
python-dotenv
google-auth
requests
transformers
torch
sentencepiece
